ai-api-integration/requirements.txt
openai
groq
transformers
torch
google-genai
cohere
python-dotenv
requests